requirements.txt
absl-py==1.3.0
biotite==0.34.1
certifi==2022.9.24
charset-normalizer==2.1.1
einops==0.5.0
fair-esm==1.0.2
idna==3.4
msgpack==1.0.4
networkx==2.8.7
numpy==1.23.4
requests==2.28.1
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
